Zaimplementowane ładowanie obiektów z plików przez bibliotekę ASSIMP.

+++ Brickpong/Brickpong/Game Engines/Brickpong/BrickpongBall.cpp
@@ -51,7 +51,7 @@ BrickpongBall::BrickpongBall(b2World* aworld, float aradius, b2Vec2 alinearVeloc
             {1.0f, 1.0f, 1.0f, }, // 8-I
     };
     VBO* cVBO = new VBO(color_data, 9, GL_STATIC_DRAW);
-    GLushort index_data[14][3] = {
+    GLuint index_data[14][3] = {
             {0, 1, 2, }, // EHD - left
             {0, 2, 4, }, // EDA - left
             {3, 0, 4, }, // BEA - back
