+++ Brickpong/Brickpong/Game Engines/Brickpong/BrickpongBrick.cpp
@@ -44,7 +44,7 @@ BrickpongBrick::BrickpongBrick(b2World* aworld, b2Vec2 aposition, float awidth, 
             {0.0f, 0.0f, 0.0f, }, // 8-I
     };
     VBO* cVBO = new VBO(color_data, 9, GL_STATIC_DRAW);
-    GLushort index_data[14][3] = {
+    GLuint index_data[14][3] = {
             {0, 1, 2, }, // EHD - left
             {0, 2, 4, }, // EDA - left
             {3, 0, 4, }, // BEA - back
